all source files use a common copyright comment header & adjusted author info [minor]
The author info (@author tag in class java-doc comments) is meant to be
used in case of questions about the code, and is currently split among
Betalord and hoijui.

The name and email in the copyright header is more a legal thing,
and is now pointing to hoijui in all files. If anyone (likely Betalord)
wants this changed, tell me please.
The FSF recommends having a real name and an email, and i do not have
that from Betalord.